공지사항추가 AWS사용

+++ src/main/java/org/yeinp/animate/animate_backend/entity/MissCareAnimal.java
@@ -14,7 +14,7 @@ import lombok.*;
 public class MissCareAnimal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int mcAnimalNo;
+    private Long mcAnimalNo;
 
     private String mcBreed;
     private String mcGender;
